+++ example/lib/main.dart
@@ -1,6 +1,9 @@
 import 'package:camerawesome/camerawesome_plugin.dart';
 import 'package:flutter/material.dart';
 import 'package:unified_apple_vision/unified_apple_vision.dart';
+import 'package:unified_apple_vision_example/extension/vision_detected_object_observation.dart';
+import 'package:unified_apple_vision_example/extension/vision_recognized_text_observation.dart';
+import 'package:unified_apple_vision_example/extension/vision_rectangle_observation.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -85,145 +88,5 @@ class _CameraScreenState extends State<CameraScreen> {
         ],
       ),
     );
-  }
-}
-
-extension VisionRecognizedTextObservationEx on VisionRecognizedTextObservation {
-  Widget build() {
-    return Positioned.fill(
-      child: CustomPaint(
-        painter: _Painter(this),
-      ),
-    );
-  }
-}
-
-class _Painter extends CustomPainter {
-  final VisionRecognizedTextObservation recognizedText;
-
-  _Painter(this.recognizedText);
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    // Draw the bounding box
-    final rect = recognizedText.scale(size).reverse(Offset(0, size.height));
-    final path = Path()
-      ..moveTo(rect.topLeft.dx, rect.topLeft.dy)
-      ..lineTo(rect.topRight.dx, rect.topRight.dy)
-      ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy)
-      ..lineTo(rect.bottomLeft.dx, rect.bottomLeft.dy)
-      ..lineTo(rect.topLeft.dx, rect.topLeft.dy);
-    final paint = Paint()
-      ..color = Colors.red
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 2;
-    canvas.drawPath(path, paint);
-
-    // Draw top-left corner
-    final cornerPaint = Paint()..style = PaintingStyle.fill;
-    canvas.drawCircle(rect.topLeft, 3, cornerPaint..color = Colors.red);
-
-    // Draw the text
-    final candidate = recognizedText.candidates.first;
-    final text = candidate.text;
-    final confidence = candidate.confidence;
-    final textPainter = TextPainter(
-      text: TextSpan(
-        text: '$text (${confidence.toStringAsFixed(1)})',
-        style: const TextStyle(fontSize: 10),
-      ),
-      textDirection: TextDirection.ltr,
-    )..layout();
-    textPainter.paint(canvas, rect.topLeft);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return false;
-  }
-}
-
-extension VisionRectangleEx on VisionRectangleObservation {
-  VisionRectangleObservation reverse(Offset offset) {
-    abs(Offset offset) => Offset(offset.dx.abs(), offset.dy.abs());
-
-    return copyWith(
-      topLeft: abs(offset - topLeft),
-      topRight: abs(offset - topRight),
-      bottomLeft: abs(offset - bottomLeft),
-      bottomRight: abs(offset - bottomRight),
-    );
-  }
-
-  Widget build() {
-    return Positioned.fill(
-      child: CustomPaint(
-        painter: _RectanglePainter(this),
-      ),
-    );
-  }
-}
-
-class _RectanglePainter extends CustomPainter {
-  final VisionRectangleObservation rectangle;
-
-  _RectanglePainter(this.rectangle);
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final rect = rectangle.scale(size).reverse(Offset(0, size.height));
-    final path = Path()
-      ..moveTo(rect.topLeft.dx, rect.topLeft.dy)
-      ..lineTo(rect.topRight.dx, rect.topRight.dy)
-      ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy)
-      ..lineTo(rect.bottomLeft.dx, rect.bottomLeft.dy)
-      ..lineTo(rect.topLeft.dx, rect.topLeft.dy);
-    final paint = Paint()
-      ..color = Colors.green
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 2;
-    canvas.drawPath(path, paint);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return false;
-  }
-}
-
-extension VisionDetectedObjectObservationEx on VisionDetectedObjectObservation {
-  Widget build() {
-    return Positioned.fill(
-      child: CustomPaint(
-        painter: _ObjectPainter(this),
-      ),
-    );
-  }
-}
-
-class _ObjectPainter extends CustomPainter {
-  final VisionDetectedObjectObservation object;
-
-  _ObjectPainter(this.object);
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final rect = object.boundingBox;
-    final path = Path()
-      ..moveTo(rect.topLeft.dx, rect.topLeft.dy)
-      ..lineTo(rect.topRight.dx, rect.topRight.dy)
-      ..lineTo(rect.bottomRight.dx, rect.bottomRight.dy)
-      ..lineTo(rect.bottomLeft.dx, rect.bottomLeft.dy)
-      ..lineTo(rect.topLeft.dx, rect.topLeft.dy);
-    final paint = Paint()
-      ..color = Colors.blue
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 2;
-    canvas.drawPath(path, paint);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return false;
   }
 }
